Avoid array allocation on .emit()
This was resulting on allocating an array with each emit.
Removing this allocation results in a ~1.75x speedup for calls
to emit, in exchange for slightly slower calls to removeListener.

Note that now the event arrays are copied when removing a listener,
to prevent unexpected behaviour when a listener is removed halfway
through the emit loop.